# docs/HOMEZ_ONCHANNEL_OPENAPI_SPEC_20260908.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "온채널 OpenAPI",
        "version": "0.1",
        "x-name": "openapi"
    },
    "paths": {
...
        "/openapi/seller/order/{code}": {
            "get": {
                "tags": [
                    "판매사 - 주문"
                ],
                "summary": "주문 상세",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "주문번호",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "GO_1011779536"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/success"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "order_code": {
                                                            "description": "주문코드",
                                                            "type": "string",
                                                            "example": "MO_1011780094"
                                                        },
                                                        "product_code": {
                                                            "description": "상품코드",
                                                            "type": "string",
                                                            "example": "CH123211"
                                                        },
                                                        "product_name": {
                                                            "description": "상품명",
                                                            "type": "string",
                                                            "example": "셀링온 쿠션퍼프 7P"
                                                        },
                                                        "order_price": {
                                                            "description": "총 주문 금액",
                                                            "type": "integer",
                                                            "example": 3120
                                                        },
                                                        "sum_product_price": {
                                                            "description": "상품비 합계",
                                                            "type": "integer",
                                                            "example": 3120
                                                        },
                                                        "sum_delivery_price": {
                                                            "description": "배송비 합계",
                                                            "type": "integer",
                                                            "example": 3000
                                                        },
                                                        "sum_add_price": {
                                                            "description": "추가 비용 합계",
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "options": {
                                                            "description": "주문 옵션",
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "description": "옵션ID",
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "description": "옵션명",
                                                                        "type": "string",
                                                                        "example": "레드 R"
                                                                    },
                                                                    "price": {
                                                                        "description": "옵션금액",
                                                                        "type": "integer",
                                                                        "example": 550
                                                                    },
                                                                    "qty": {
                                                                        "description": "수량",
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "status": {
                                                            "$ref": "#/components/schemas/order.status"
                                                        },
                                                        "detail_status": {
                                                            "description": "상세 주문상태",
                                                            "type": "string",
                                                            "example": "발송완료"
                                                        },
                                                        "recv_info": {
                                                            "description": "수령인 정보",
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "recv_name": {
                                                                        "description": "수령인명",
                                                                        "type": "string",
                                                                        "example": "김수현"
                                                                    },
                                                                    "recv_tell": {
                                                                        "description": "수령인 전화번호",
                                                                        "type": "string",
                                                                        "example": "010-0000-0000"
                                                                    },
                                                                    "recv_mobile": {
                                                                        "description": "수령인 휴대폰번호",
                                                                        "type": "string",
                                                                        "example": "010-0000-0000"
                                                                    },
                                                                    "zipcode": {
                                                                        "description": "우편번호",
                                                                        "type": "string",
                                                                        "example": "04782"
                                                                    },
                                                                    "address": {
                                                                        "description": "주소",
                                                                        "type": "string",
                                                                        "example": "서울특별시 성동구 연무장1길 13"
                                                                    },
                                                                    "address_detail": {
                                                                        "description": "상세주소",
                                                                        "type": "string",
                                                                        "example": ""
                                                                    },
                                                                    "comments": {
                                                                        "description": "옵션별 메세지",
                                                                        "type": "array",
                                                                        "items": {
                                                                            "properties": {
                                                                                "option_id": {
                                                                                    "description": "옵션ID",
                                                                                    "type": "integer",
                                                                                    "example": 1
                                                                                },
                                                                                "option_name": {
                                                                                    "description": "옵션명",
                                                                                    "type": "string",
                                                                                    "example": "레드 R"
                                                                                },
                                                                                "comment": {
                                                                                    "description": "메세지",
                                                                                    "type": "string",
                                                                                    "example": "안전배송해주세요."
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        }
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "deliverys": {
                                                            "description": "배송 정보",
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "name": {
                                                                        "description": "택배사명",
                                                                        "type": "string",
                                                                        "example": "CJ 대한통운"
                                                                    },
                                                                    "tracking_number": {
                                                                        "description": "운송장번호",
                                                                        "type": "string",
                                                                        "example": "696870140211"
                                                                    },
                                                                    "created_at": {
                                                                        "description": "등록일",
                                                                        "type": "string",
                                                                        "example": "2026-03-13 11:56:19"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "created_at": {
                                                            "description": "등록일",
                                                            "type": "string",
                                                            "example": "2026-03-13 11:54:39"
                                                        },
                                                        "update_at": {
                                                            "description": "수정일",
                                                            "type": "string",
                                                            "example": "2026-03-13 11:56:19"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "success": {
                "title": "통신 성공 Scheme",
                "description": "통신 성공 Scheme",
                "required": [
                    "status",
                    "meta",
                    "result"
                ],
                "properties": {
                    "status": {
                        "description": "상태코드",
                        "type": "integer",
                        "example": 200
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "description": "통신 시간",
                                "type": "string",
                                "example": "2026-01-01 00:00:00"
                            },
                            "api_type": {
                                "description": "api 종류",
                                "type": "string",
                                "example": "openapi"
                            }
                        },
                        "type": "object"
                    },
                    "result": {
                        "description": "결과",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "error": {
                "title": "통신 실패 Scheme",
                "description": "통신 실패 Scheme",
                "required": [
                    "status",
                    "meta",
                    "error"
                ],
                "properties": {
                    "status": {
                        "description": "상태코드",
                        "type": "integer",
                        "example": 401
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "description": "통신 시간",
                                "type": "string",
                                "example": "2026-01-01 00:00:00"
                            },
                            "api_type": {
                                "description": "api 종류",
                                "type": "string",
                                "example": "openapi"
                            }
                        },
                        "type": "object"
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "description": "에러코드",
                                "type": "integer",
                                "example": 401
                            },
                            "message": {
                                "description": "에러 내용",
                                "type": "string",
                                "example": "잘못 된 접근입니다."
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "order.status": {
                "title": "주문단계",
                "description": "1: 입금확인, 2: 배송준비중, 3: 배송완료, 4: 취소반품, 5: 부분환불",
                "type": "integer",
                "enum": [
                    1,
                    2,
                    3,
                    4,
                    5
                ],
                "example": 2
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}